requirements.txt
Pillow
pyTelegramBotAPI
uuid
